[spike] custom mirage serializer and class builder
This spike was done to gain an understanding of how we could utilize
ember-cli-mirage.

1. Can we use Mirage to build test fixtures
We determined that using the mirage server, we are able to hydrate any
model, not just EmberData (see SimpleModel test), for use in
integration and unit tests.

2. Maintaining Serializers seems like a lot of overhead compared to
FactoryGuy
We learned that using the `makeServer()` config hook and
`applyEmberDataSerializers` helper in order to automate building our
Mirage Serializers from our Ember Data serializers. This makes it
easier to maintain Mirage Serializers since you only need to be
explicit about including sideloaded data, ie data which is not embedded.
https://github.com/miragejs/ember-cli-mirage/blob/324a0322a63949f7af1fddf61ed29c90b16249f9/addon/serializers/ember-data-serializer.js#L21-L25